Move `env` subcommand to `internal-tools`
From `gix-testtools`, where it was originally placed.

This also adapts and improves the original description from the
commit message that had introduced the subcommand, to make clear
when it makes sense to use this (given that operating systems and
shells already supply facilities for listing environment variables).
