+++ tests/it/src/commands/env.rs
@@ -1,0 +1,15 @@
+pub(super) mod function {
+    pub fn env() -> anyhow::Result<()> {
+        for (name, value) in std::env::vars_os() {
+            println!("{}={}", repr(&name), repr(&value));
+        }
+        Ok(())
+    }
+
+    fn repr(text: &std::ffi::OsStr) -> String {
+        text.to_str()
+            .filter(|s| !s.contains('"'))
+            .map(ToOwned::to_owned)
+            .unwrap_or_else(|| format!("{text:?}"))
+    }
+}
