+++ tests/it/src/commands/mod.rs
@@ -6,3 +6,6 @@ pub use git_to_sh::function::git_to_sh;
 
 pub mod check_mode;
 pub use check_mode::function::check_mode;
+
+pub mod env;
+pub use env::function::env;
